fix: moving footbar to common folder
