+++ apps/frontend/src/pages/index.tsx
@@ -1,7 +1,7 @@
 import About from "@/components/LandingPage/About";
 import Acheivement from "@/components/LandingPage/Acheivement";
 import Feature from "@/components/LandingPage/Feature";
-import FooterBar from "@/components/LandingPage/FooterBar";
+import FooterBar from "@/components/common/FooterBar";
 import Hero from "@/components/LandingPage/Hero";
 import { Grid } from "@mui/material";
 import React from "react";
